+++ client/Brain.tsx
@@ -4,6 +4,14 @@ import type { APIResponse } from "../global";
 import { APIVersion } from "../version";
 import { durationAgo } from "./utils";
 
+const updateInterval = 60_000; // One minute in milliseconds
+const refreshInterval = 3600_000 + 60_000; // One hour and one minute in milliseconds
+
+// const fail = async (ignore: any) => {
+//   await new Promise((resolve) => setTimeout(resolve, 1_000)); // 1 sec
+//   return new Response(null, { status: 200 });
+// };
+
 export const Brain: FC = () => {
   const { globalStatus, setGlobalStatus, setGlobalData, setGlobalClient, setError } =
     useContext(GlobalContext);
@@ -11,18 +19,27 @@ export const Brain: FC = () => {
 
   const fetchAPI = async (endpoint: string): Promise<APIResponse> => {
     try {
+      // const response = await fail(endpoint);
       const response = await fetch(endpoint);
-      const data: APIResponse = await response.json();
-      if (!data) {
-        throw "No data received from API";
+      if (!response.ok) {
+        throw `Unsuccessful response status ${response.status} received`;
+      }
+      const responseText = await response.text();
+      if (responseText.length === 0) {
+        throw "Empty response";
+      }
+      const data: APIResponse = JSON.parse(responseText);
+      if (!data || Object.keys(data).length === 0) {
+        throw "No JSON data received";
       }
       if (data.version !== APIVersion) {
-        throw "API version mismatch: Please try reloading, clear caches, etc.";
+        throw "Version mismatch: Please try reloading, clear caches, etc.";
       }
       return data;
-    } catch (err: any) {
-      setError(err ?? "API error", true);
-      console.error(err);
+    } catch (err) {
+      const errorMessage = `API failed to load, please try again in a few minutes (${err})`;
+      setError(errorMessage, true);
+      console.error(errorMessage);
     }
     return { version: -1 };
   };
@@ -43,8 +60,9 @@ export const Brain: FC = () => {
     loadAPI()
       .then(() => setStartIntervalTrigger(true))
       .catch((err) => {
-        setError(err ?? "Error loading data from API", true);
-        console.error(err);
+        const errorMessage = `Error loading data from API: ${err}`;
+        setError(errorMessage, true);
+        console.error(errorMessage);
       });
   }, []);
 
@@ -55,7 +73,7 @@ export const Brain: FC = () => {
     const handleHardRefresh = async () => {
       // Only load data from the API when database has changed
       try {
-        const prevDbTimestamp = new Date(globalStatus.dbLastChange).getTime();
+        const prevDbTimestamp = new Date(localStatus.dbLastChange).getTime();
         const { status } = await fetchAPI("/api/status");
         if (status) {
           setGlobalStatus(() => status);
@@ -69,13 +87,14 @@ export const Brain: FC = () => {
           }
         }
       } catch (err: any) {
-        setError(err ?? "Error reloading data from API", true);
-        console.error(err);
+        const errorMessage = `Error reloading data from API: ${err}`;
+        setError(errorMessage, true);
+        console.error(errorMessage);
       }
       if (interval) {
         clearInterval(interval); // Kill existing interval
         updateLoop(); // Show new duration immediately because interval starts with delay
-        interval = setInterval(updateLoop, 60_000); // Restart update every minute
+        interval = setInterval(updateLoop, updateInterval); // Restart update every minute
       }
     };
 
@@ -85,21 +104,26 @@ export const Brain: FC = () => {
         ...prevState,
         navBarDurations: {
           dbLastChange: durationAgo(localStatus.dbLastChange),
-          apiLastCheck: localStatus.isDevelopment ? "[dev mode]" : durationAgo(localStatus.apiLastCheck, true),
-        }
-       }));
+          apiLastCheck: localStatus.isDevelopment
+            ? "[dev mode]"
+            : durationAgo(localStatus.apiLastCheck, true),
+        },
+      }));
       // If last API check is longer than an hour ago, check for new data, trigger a refresh if needed
       const now = Date.now();
-      const oneHourAndOneMinute = 3600_000 + 60_000; // 1 hour + 1 minute in milliseconds
-      if (now - new Date(localStatus.apiLastCheck).getTime() > oneHourAndOneMinute) {
-        handleHardRefresh().catch(console.error);
+      if (now - new Date(localStatus.apiLastCheck).getTime() > refreshInterval) {
+        handleHardRefresh().catch((err: any) => {
+          const errorMessage = `Error trying to reload data from API: ${err}`;
+          setError(errorMessage, true);
+          console.error(errorMessage);
+        });
       }
     };
 
     // Start the 60 second update interval if API data is valid, otherwise load data from API
     if (localStatus.isValid && !interval) {
       updateLoop(); // Immediate call to update duration strings because interval starts with delay
-      interval = setInterval(updateLoop, 60_000); // Update every minute
+      interval = setInterval(updateLoop, updateInterval); // Update every minute
     }
 
     // Clean up the interval when the component is unmounted
@@ -110,6 +134,5 @@ export const Brain: FC = () => {
     };
   }, [startIntervalTrigger]);
 
-
-  return <></>
-}
+  return <></>;
+};